pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dino-peft"
version = "0.0.1"
description = "LoRA/PEFT on DINOv2 for EM segmentation"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "numpy",
  "Pillow",
  "pyyaml",
  "open-clip-torch",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
